+++ ams/opt/constraint.py
@@ -64,6 +64,12 @@ class Constraint(OptzBase):
         self.dual = None
         self.code = None
 
+    def get_idx(self):
+        raise NotImplementedError
+
+    def get_all_idxes(self):
+        raise NotImplementedError
+
     @ensure_symbols
     def parse(self):
         """