+++ docs/source/release-notes.rst
@@ -13,7 +13,7 @@ v0.9.14 (2024-xx-xx)
 --------------------
 
 - **Breaking Change**: rename model ``Region`` to ``Zone`` for clarity. Prior case
-  files can run into error.
+  files without modification can run into error.
 
 v0.9.13 (2024-12-05)
 --------------------