+++ docs/source/release-notes.rst
@@ -9,6 +9,14 @@ The APIs before v3.0.0 are in beta and may change without prior notice.
 Pre-v1.0.0
 ==========
 
+v0.9.14 (2024-xx-xx)
+--------------------
+
+- **Breaking Change**: rename model ``Region`` to ``Zone`` for clarity. Prior case
+  files without modification can run into error.
+- Fix bugs in ``RTED.dc2ac``
+- Minor refacot ``OptzBase.get_idx`` to reduce duplication
+
 v0.9.13 (2024-12-05)
 --------------------